+++ dashboard-app/src/dashboard/components/product-drawer/secondary/model/secondaryConfirmedRoundModel.ts
@@ -1,6 +1,6 @@
 export interface SecondaryConfirmedRound {
   date: string
-  excludePeriodExistingOrderInbound: boolean
+  excludeSegmentExistingOrderInbound: boolean
   qtyBySize: Record<string, number>
 }